Reconstruct the code to avoid mcp tools imported from wrapper function in tool_wrapper.py
